refactor: remove subfolders for store
