+++ ts/features/it-wallet/saga/index.ts
@@ -1,13 +1,13 @@
 import { fork } from "typed-redux-saga/macro";
 import { SagaIterator } from "redux-saga";
 
-import { watchItwIssuancePidSaga } from "./issuance/pid/itwIssuancePidSaga";
+import { watchItwIssuancePidSaga } from "./itwIssuancePidSaga";
 import { watchItwWiaSaga } from "./itwWiaSaga";
 import { watchItwActivationSaga } from "./itwActivationSaga";
-import { watchitwPrRemoteCredentialSaga } from "./presentation/remote/itwPrRemoteCredentialSaga";
-import { watchItwIssuanceCredentialSaga } from "./issuance/credential/itwIssuanceCredentialSaga";
-import { watchItwPrProximitySaga } from "./presentation/proximity/itwPrProximitySaga";
-import { watchItwPrRemotePid } from "./presentation/remote/itwPrRemotePidSaga";
+import { watchitwPrRemoteCredentialSaga } from "./itwPrRemoteCredentialSaga";
+import { watchItwIssuanceCredentialSaga } from "./itwIssuanceCredentialSaga";
+import { watchItwPrProximitySaga } from "./itwPrProximitySaga";
+import { watchItwPrRemotePid } from "./itwPrRemotePidSaga";
 
 /**
  * Watcher for any IT wallet related sagas.
